lib/h2ph.t: keep tmpfiles on error
It's extremely hard to analyse h2ph errors without those tempfiles,
and errors do not happen for the normal case.

+++ lib/h2ph.t
@@ -57,6 +57,8 @@ is( $result, '', "output free of warnings" );
 
 # cleanup
 END {
+  if (!$result) {
     1 while unlink("lib/h2ph.ph");
     1 while unlink("_h2ph_pre.ph");
+  }
 }
